299. Validações parte 1 - Controle de fluxos

+++ app/Http/Controllers/MarcaController.php
@@ -40,6 +40,9 @@ public function __construct(Marca $marca) {
     public function show( $id )
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -52,11 +55,11 @@ public function __construct(Marca $marca) {
      */
     public function update(Request $request, $id)
     {
-        // print_r($request->all()); // dados atualizados
-        // echo   '<hr>';
-        // print_r($marca->getAttributes()); // dados antigos
 
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a atualização, o recurso pesquisado não existe'];
+        }
         $marca->update($request->all());
 
         return $marca;
@@ -70,6 +73,9 @@ public function __construct(Marca $marca) {
     {
         // print_r($marca->getAttributes());
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Impossível deletar o registro, o recurso pesquisado não existe'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
